Added Classes & set up console data

+++ start_code/db/imdb.sql
@@ -17,5 +17,5 @@ CREATE TABLE castings(
   id SERIAL8 PRIMARY KEY,
   movie_id INT8 REFERENCES movies(id) ON DELETE CASCADE,
   star_id INT8 REFERENCES stars(id) ON DELETE CASCADE,
-  fee INT8
+  fee INT
 );
